+++ roop/core_pipeline.py
@@ -2,12 +2,11 @@ import argparse
 import os
 from pyparsing import Any
 
-from roop.pipeline_processing.pipeline import *
-from roop.pipeline_processing.extractors.basic_extractor import BasicExtractor
-from roop.pipeline_processing.collectors.basic_collector import BasicCollector
-from roop.pipeline_processing.blocks.dummy import Dummy
-from roop.pipeline_processing.blocks.swapper import Swapper
-from roop.pipeline_processing.postprocessors.basic_postprocess import BasicPostprocess
+from roop.pipeline_processing import PipelineEnvironment, Pipeline, PipelineExecutionParams, CancellationToken
+from roop.pipeline_processing.extractors import BasicExtractor
+from roop.pipeline_processing.collectors import BasicCollector
+from roop.pipeline_processing.process_blocks import Dummy, Swapper
+from roop.pipeline_processing.postprocessors import BasicPostprocess
 
 parser = argparse.ArgumentParser()
 parser.add_argument('-f', '--face', help='use this face', dest='source_img')
@@ -18,13 +17,14 @@ args = {}
 for name, value in vars(parser.parse_args()).items():
     args[name] = value
 
+
 def create_pipeline():
     environment = PipelineEnvironment(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
     pipeline = Pipeline(
         BasicExtractor(environment),
         [
-            #Dummy()
+            #Dummy(),
             Swapper(environment)
         ],
         BasicCollector(environment),
@@ -32,15 +32,21 @@ def create_pipeline():
     )
     return pipeline
 
+
 def progress_handler(value: str):
     #print(value)
     return
 
+
 def preview_handler(frame: Any):
     return
 
+
 def run():
     pipeline = create_pipeline()
+
+    if not pipeline:
+        return
 
     params = PipelineExecutionParams(
         args['source_img'],
@@ -52,6 +58,7 @@ def run():
     token = CancellationToken()
 
     pipeline.execute(params, token)
+
 
 if __name__ == '__main__':
     run()
