+++ roop/pipeline_processing/cancellation_token.py
@@ -1,0 +1,9 @@
+class CancellationToken:
+    """Cancellation token"""
+
+    def __init__(self):
+        self.cancelled = False
+
+    def cancel(self):
+        """Set cancelled"""
+        self.cancelled = True
