+++ roop/pipeline_processing/collectors/__init__.py
@@ -1,0 +1,7 @@
+from .frame_collector import FramesCollector
+from .basic_collector import BasicCollector
+
+__all__ = [
+    'FramesCollector',
+    'BasicCollector',    
+]
