+++ roop/pipeline_processing/extractors/__init__.py
@@ -1,0 +1,7 @@
+from .frame_extractor import FramesExtractor
+from .basic_extractor import BasicExtractor
+
+__all__ = [
+    'FramesExtractor',
+    'BasicExtractor',    
+]
