+++ roop/pipeline_processing/process_blocks/__init__.py
@@ -1,0 +1,9 @@
+from .process_block import ProcessBlock
+from .dummy import Dummy
+from .swapper import Swapper
+
+__all__ = [
+    'ProcessBlock',
+    'Dummy',
+    'Swapper',
+]
